+++ apps/prairielearn/src/lib/assessment.ts
@@ -6,8 +6,7 @@ import { z } from 'zod';
 import { callbackify, promisify } from 'util';
 
 import * as error from '@prairielearn/error';
-import { gradeVariant } from './grading';
-import * as externalGrader from './externalGrader';
+import { gradeVariantAsync } from './grading';
 import * as sqldb from '@prairielearn/postgres';
 import * as ltiOutcomes from './ltiOutcomes';
 import { createServerJob } from './server-jobs';
@@ -35,12 +34,13 @@ export const InstanceLogSchema = z.object({
   submission_id: z.string().nullable(),
   data: z.record(z.any()).nullable(),
   client_fingerprint: ClientFingerprintSchema.nullable(),
+  client_fingerprint_number: z.number().nullable(),
   formatted_date: z.string(),
   date_iso8601: z.string(),
   student_question_number: z.string().nullable(),
   instructor_question_number: z.string().nullable(),
 });
-type InstanceLogEntry = z.infer<typeof InstanceLogSchema> & { client_fingerprint_number?: number };
+export type InstanceLogEntry = z.infer<typeof InstanceLogSchema>;
 
 /**
  * Check that an assessment_instance_id really belongs to the given assessment_id
@@ -108,6 +108,7 @@ export async function makeAssessmentInstance(
   mode: 'Exam' | 'Homework',
   time_limit_min: number | null,
   date: Date,
+  client_fingerprint_id: string | null,
 ): Promise<string> {
   const result = await sqldb.callOneRowAsync('assessment_instances_insert', [
     assessment_id,
@@ -117,6 +118,7 @@ export async function makeAssessmentInstance(
     mode,
     time_limit_min,
     date,
+    client_fingerprint_id,
   ]);
   return result.rows[0].assessment_instance_id;
 }
@@ -179,13 +181,10 @@ export async function gradeAssessmentInstanceAsync(
   requireOpen: boolean,
   close: boolean,
   overrideGradeRate: boolean,
+  client_fingerprint_id: string | null,
 ): Promise<void> {
   debug('gradeAssessmentInstance()');
   overrideGradeRate = close || overrideGradeRate;
-
-  // We may have to submit grading jobs to the external grader after this
-  // grading transaction has been accepted; collect those job ids here.
-  const externalGradingJobIds: string[] = [];
 
   if (requireOpen || close) {
     await sqldb.runInTransactionAsync(async () => {
@@ -199,6 +198,7 @@ export async function gradeAssessmentInstanceAsync(
         await sqldb.queryAsync(sql.close_assessment_instance, {
           assessment_instance_id,
           authn_user_id,
+          client_fingerprint_id,
         });
       }
     });
@@ -213,7 +213,7 @@ export async function gradeAssessmentInstanceAsync(
   await async.eachSeries(variants, async (row) => {
     debug('gradeAssessmentInstance()', 'loop', 'variant.id:', row.variant.id);
     const check_submission_id = null;
-    const gradingJobId = await promisify(gradeVariant)(
+    await gradeVariantAsync(
       row.variant,
       check_submission_id,
       row.question,
@@ -221,14 +221,7 @@ export async function gradeAssessmentInstanceAsync(
       authn_user_id,
       overrideGradeRate,
     );
-    if (gradingJobId !== undefined) {
-      externalGradingJobIds.push(gradingJobId);
-    }
   });
-  if (externalGradingJobIds.length > 0) {
-    // We need to submit these grading jobs to be graded
-    await externalGrader.beginGradingJobs(externalGradingJobIds);
-  }
   // The `grading_needed` flag was set by the closing query above. Once we've
   // successfully graded every part of the assessment instance, set the flag to
   // false so that we don't try to grade it again in the future.
@@ -313,6 +306,7 @@ export async function gradeAllAssessmentInstances(
         requireOpen,
         close,
         overrideGradeRate,
+        null,
       );
     });
   });
@@ -357,6 +351,48 @@ export async function updateAssessmentStatistics(assessment_id: string): Promise
 
     // update the statistics
     await sqldb.queryOneRowAsync(sql.update_assessment_statisics, { assessment_id });
+  });
+}
+
+export async function updateAssessmentInstanceScore(
+  assessment_instance_id: string,
+  score_perc: number,
+  authn_user_id: string,
+): Promise<void> {
+  await sqldb.runInTransactionAsync(async () => {
+    const max_points = await sqldb.queryRow(
+      sql.select_and_lock_assessment_instance_max_points,
+      { assessment_instance_id },
+      z.number(),
+    );
+    const points = (score_perc * max_points) / 100;
+    await sqldb.queryAsync(sql.update_assessment_instance_score, {
+      assessment_instance_id,
+      score_perc,
+      points,
+      authn_user_id,
+    });
+  });
+}
+
+export async function updateAssessmentInstancePoints(
+  assessment_instance_id: string,
+  points: number,
+  authn_user_id: string,
+): Promise<void> {
+  await sqldb.runInTransactionAsync(async () => {
+    const max_points = await sqldb.queryRow(
+      sql.select_and_lock_assessment_instance_max_points,
+      { assessment_instance_id },
+      z.number(),
+    );
+    const score_perc = (points / (max_points > 0 ? max_points : 1)) * 100;
+    await sqldb.queryAsync(sql.update_assessment_instance_score, {
+      assessment_instance_id,
+      score_perc,
+      points,
+      authn_user_id,
+    });
   });
 }