+++ apps/prairielearn/src/lib/groups.ts
@@ -1,36 +1,29 @@
 import * as error from '@prairielearn/error';
-import { flash } from '@prairielearn/flash';
 import { z } from 'zod';
 import _ = require('lodash');
 
 import * as sqldb from '@prairielearn/postgres';
 import { idsEqual } from './id';
-import { IdSchema } from './db-types';
+import {
+  GroupSchema,
+  IdSchema,
+  type User,
+  UserSchema,
+  GroupConfigSchema,
+  type GroupConfig,
+  GroupRoleSchema,
+  type GroupUserRole,
+} from './db-types';
+import { getEnrollmentForUserInCourseInstance } from '../models/enrollment';
+import { selectUserByUid } from '../models/user';
 const sql = sqldb.loadSqlEquiv(__filename);
 
-const GroupConfigSchema = z.object({
-  assessment_id: z.string().nullable(),
-  course_instance_id: z.string(),
-  date: z.date().nullable(),
-  deleted_at: z.date().nullable(),
-  has_roles: z.boolean().nullable(),
-  id: z.string(),
-  maximum: z.number().nullable(),
-  minimum: z.number().nullable(),
-  name: z.string().nullable(),
-  student_authz_create: z.boolean().nullable(),
-  student_authz_join: z.boolean().nullable(),
-  student_authz_leave: z.boolean().nullable(),
-});
-type GroupConfig = z.infer<typeof GroupConfigSchema>;
-
-const GroupMemberSchema = z.object({
-  uid: z.string(),
-  user_id: z.string(),
-  group_name: z.string(),
-  join_code: z.string(),
-});
-type GroupMember = z.infer<typeof GroupMemberSchema>;
+export class GroupOperationError extends Error {
+  constructor(message) {
+    super(message);
+    this.name = 'GroupOperationError';
+  }
+}
 
 const RoleAssignmentSchema = z.object({
   user_id: z.string(),
@@ -40,34 +33,22 @@ const RoleAssignmentSchema = z.object({
 });
 type RoleAssignment = z.infer<typeof RoleAssignmentSchema>;
 
-const GroupRoleSchema = z.object({
-  id: z.string(),
-  role_name: z.string(),
-  count: z.coerce.number(),
-  maximum: z.number().nullable(),
-  minimum: z.number().nullable(),
-  can_assign_roles_at_start: z.boolean(),
-  can_assign_roles_during_assessment: z.boolean(),
+const GroupRoleWithCountSchema = GroupRoleSchema.extend({
+  count: z.number(),
 });
-type GroupRole = z.infer<typeof GroupRoleSchema>;
-
-const AssessmentLevelPermissionsSchema = z.object({
-  can_assign_roles_at_start: z.boolean().nullable(),
-  can_assign_roles_during_assessment: z.boolean().nullable(),
-});
-type AssessmentLevelPermissions = z.infer<typeof AssessmentLevelPermissionsSchema>;
+type GroupRoleWithCount = z.infer<typeof GroupRoleWithCountSchema>;
 
 interface RolesInfo {
   roleAssignments: Record<string, RoleAssignment[]>;
-  groupRoles: GroupRole[];
-  validationErrors: GroupRole[];
+  groupRoles: GroupRoleWithCount[];
+  validationErrors: GroupRoleWithCount[];
   disabledRoles: string[];
   rolesAreBalanced: boolean;
-  usersWithoutRoles: GroupMember[];
+  usersWithoutRoles: User[];
 }
 
 interface GroupInfo {
-  groupMembers: GroupMember[];
+  groupMembers: User[];
   groupSize: number;
   groupName: string;
   joinCode: string;
@@ -75,10 +56,13 @@ interface GroupInfo {
   rolesInfo?: RolesInfo;
 }
 
-interface GroupRoleAssignment {
-  group_role_id: string;
-  user_id: string;
-}
+type GroupRoleAssignment = Pick<GroupUserRole, 'group_role_id' | 'user_id'>;
+
+const GroupForUpdateSchema = GroupSchema.extend({
+  cur_size: z.number(),
+  max_size: z.number().nullable(),
+  has_roles: z.boolean(),
+});
 
 /**
  * Gets the group config info for a given assessment id.
@@ -103,23 +87,21 @@ export async function getGroupId(assessmentId: string, userId: string): Promise<
   );
 }
 
-export async function getGroupInfo(groupId: string, groupConfig: GroupConfig): Promise<GroupInfo> {
-  const result = await sqldb.queryRows(
-    sql.get_group_members,
-    { group_id: groupId },
-    GroupMemberSchema,
-  );
-  const needSize = (groupConfig.minimum ?? 0) - result.length;
+export async function getGroupInfo(group_id: string, groupConfig: GroupConfig): Promise<GroupInfo> {
+  const group = await sqldb.queryRow(sql.select_group, { group_id }, GroupSchema);
+  const groupMembers = await sqldb.queryRows(sql.select_group_members, { group_id }, UserSchema);
+
+  const needSize = (groupConfig.minimum ?? 0) - groupMembers.length;
   const groupInfo: GroupInfo = {
-    groupMembers: result,
-    groupSize: result.length,
-    groupName: result[0].group_name,
-    joinCode: result[0].group_name + '-' + result[0].join_code,
+    groupMembers,
+    groupSize: groupMembers.length,
+    groupName: group.name,
+    joinCode: group.name + '-' + group.join_code,
     start: needSize <= 0,
   };
 
   if (groupConfig.has_roles) {
-    const rolesInfo = await getRolesInfo(groupId, groupInfo.groupMembers);
+    const rolesInfo = await getRolesInfo(group_id, groupInfo.groupMembers);
     groupInfo.start =
       groupInfo.start &&
       rolesInfo.rolesAreBalanced &&
@@ -135,7 +117,7 @@ export async function getGroupInfo(groupId: string, groupConfig: GroupConfig): P
  * A helper function to getGroupInfo that returns a data structure containing info about an
  * assessment's group roles.
  */
-async function getRolesInfo(groupId: string, groupMembers: GroupMember[]): Promise<RolesInfo> {
+async function getRolesInfo(groupId: string, groupMembers: User[]): Promise<RolesInfo> {
   // Get the current role assignments of the group
   const result = await sqldb.queryRows(
     sql.get_role_assignments,
@@ -148,7 +130,7 @@ async function getRolesInfo(groupId: string, groupMembers: GroupMember[]): Promi
   const groupRoles = await sqldb.queryRows(
     sql.get_group_roles,
     { group_id: groupId },
-    GroupRoleSchema,
+    GroupRoleWithCountSchema,
   );
 
   // Identify errors for any roles where count is not between max and min (if they exist)
@@ -184,86 +166,226 @@ async function getRolesInfo(groupId: string, groupMembers: GroupMember[]): Promi
   };
 }
 
+/**
+ * This function assumes that the group has roles, so any caller must ensure
+ * that it is only called in that scenario
+ */
+export async function getQuestionGroupPermissions(
+  instance_question_id: string,
+  group_id: string,
+  user_id: string,
+): Promise<{ can_submit: boolean; can_view: boolean }> {
+  const userPermissions = await sqldb.queryOptionalRow(
+    sql.select_question_permissions,
+    { instance_question_id, group_id, user_id },
+    z.object({ can_submit: z.boolean(), can_view: z.boolean() }),
+  );
+  return userPermissions ?? { can_submit: false, can_view: false };
+}
+
+export async function addUserToGroup({
+  assessment_id,
+  group_id,
+  uid,
+  authn_user_id,
+  enforceGroupSize,
+}: {
+  assessment_id: string;
+  group_id: string;
+  uid: string;
+  authn_user_id: string;
+  enforceGroupSize: boolean;
+}) {
+  await sqldb.runInTransactionAsync(async () => {
+    const group = await sqldb.queryOptionalRow(
+      sql.select_and_lock_group,
+      { group_id, assessment_id },
+      GroupForUpdateSchema,
+    );
+    if (group == null) {
+      throw new GroupOperationError(`Group does not exist.`);
+    }
+
+    const user = await selectUserByUid(uid);
+    const userIsInstructor =
+      user &&
+      (await sqldb.callRow(
+        'users_is_instructor_in_course_instance',
+        [user.user_id, group.course_instance_id],
+        z.boolean(),
+      ));
+    const userIsStudent =
+      user &&
+      !userIsInstructor &&
+      !!(await getEnrollmentForUserInCourseInstance({
+        course_instance_id: group.course_instance_id,
+        user_id: user.user_id,
+      }));
+    // To be part of a group, the user needs to either be enrolled in the course
+    // instance, or be an instructor
+    if (!userIsStudent && !userIsInstructor) {
+      throw new GroupOperationError(`User ${uid} is not enrolled in this course.`);
+    }
+
+    // This is technically susceptible to race conditions. That won't be an
+    // issue once we have a unique constraint for group membership.
+    const existingGroupId = await getGroupId(assessment_id, user.user_id);
+    if (existingGroupId != null) {
+      if (idsEqual(user.user_id, authn_user_id)) {
+        throw new GroupOperationError('You are already in another group.');
+      } else {
+        throw new GroupOperationError('User is already in another group.');
+      }
+    }
+
+    if (enforceGroupSize && group.max_size != null && group.cur_size >= group.max_size) {
+      throw new GroupOperationError(`Group is already full.`);
+    }
+
+    // Find a group role. If none of the roles can be assigned, assign no role.
+    const groupRoleId = group.has_roles
+      ? await sqldb.queryOptionalRow(
+          sql.select_suitable_group_role,
+          { assessment_id, group_id: group.id, cur_size: group.cur_size },
+          IdSchema,
+        )
+      : null;
+
+    await sqldb.queryAsync(sql.insert_group_user, {
+      group_id: group.id,
+      user_id: user.user_id,
+      group_config_id: group.group_config_id,
+      authn_user_id: authn_user_id,
+      group_role_id: groupRoleId,
+    });
+  });
+}
+
 export async function joinGroup(
   fullJoinCode: string,
-  assessmentId: string,
-  userId: string,
-  authnUserId: string,
+  assessment_id: string,
+  uid: string,
+  authn_user_id: string,
 ): Promise<void> {
   const splitJoinCode = fullJoinCode.split('-');
   if (splitJoinCode.length !== 2 || splitJoinCode[1].length !== 4) {
     // the join code input by user is not valid (not in format of groupname+4-character)
-    flash('error', 'The join code has an incorrect format');
+    throw new GroupOperationError('The join code has an incorrect format');
     return;
   }
 
-  const groupName = splitJoinCode[0];
-  const joinCode = splitJoinCode[1].toUpperCase();
-
-  // This is a best-effort check to produce a nice error message. Even if this
-  // fails due to a race condition, the `group_users_insert` sproc below will
-  // validate that the user isn't already in a group.
-  const existingGroupId = await getGroupId(assessmentId, userId);
-  if (existingGroupId != null) {
-    flash('error', 'You are already in another group.');
-    return;
-  }
+  const group_name = splitJoinCode[0];
+  const join_code = splitJoinCode[1].toUpperCase();
 
   try {
-    await sqldb.callAsync('group_users_insert', [
-      assessmentId,
-      userId,
-      authnUserId,
-      groupName,
-      joinCode,
-    ]);
+    await sqldb.runInTransactionAsync(async () => {
+      const group = await sqldb.queryOptionalRow(
+        sql.select_and_lock_group_by_name,
+        { group_name, assessment_id },
+        GroupSchema,
+      );
+      if (group == null || group.join_code !== join_code) {
+        throw new GroupOperationError('Group does not exist.');
+      }
+      await addUserToGroup({
+        assessment_id,
+        group_id: group.id,
+        uid,
+        authn_user_id,
+        enforceGroupSize: true,
+      });
+    });
   } catch (err) {
-    flash(
-      'error',
-      `Failed to join the group with join code ${fullJoinCode}. It is already full or does not exist. Please try to join another one.`,
-    );
+    if (err instanceof GroupOperationError) {
+      throw new GroupOperationError(`Cannot join group "${fullJoinCode}": ${err.message}`);
+    }
+    throw err;
   }
 }
 
 export async function createGroup(
-  groupName: string,
-  assessmentId: string,
-  userId: string,
-  authnUserId: string,
+  group_name: string,
+  assessment_id: string,
+  uids: string[],
+  authn_user_id: string,
 ): Promise<void> {
-  if (groupName.length > 30) {
-    flash('error', 'The group name is too long. Use at most 30 alphanumerical characters.');
-    return;
+  if (group_name.length > 30) {
+    throw new GroupOperationError(
+      'The group name is too long. Use at most 30 alphanumerical characters.',
+    );
   }
-  if (!groupName.match(/^[0-9a-zA-Z]+$/)) {
-    flash(
-      'error',
+  if (!group_name.match(/^[0-9a-zA-Z]+$/)) {
+    throw new GroupOperationError(
       'The group name is invalid. Only alphanumerical characters (letters and digits) are allowed.',
     );
-    return;
   }
 
-  // This is technically susceptible to race conditions. That won't be an
-  // issue once we have a unique constraint for group membership.
-  const existingGroupId = await getGroupId(assessmentId, userId);
-  if (existingGroupId != null) {
-    flash('error', 'You are already in a group.');
-    return;
+  if (uids.length === 0) {
+    throw new GroupOperationError('There must be at least one user in the group.');
   }
 
   try {
-    await sqldb.queryAsync(sql.create_group, {
-      assessment_id: assessmentId,
-      user_id: userId,
-      authn_user_id: authnUserId,
-      group_name: groupName,
+    await sqldb.runInTransactionAsync(async () => {
+      let group_id;
+      try {
+        group_id = await sqldb.queryRow(
+          sql.create_group,
+          { assessment_id, authn_user_id, group_name },
+          IdSchema,
+        );
+      } catch (err) {
+        // 23505 is the Postgres error code for unique constraint violation
+        // (https://www.postgresql.org/docs/current/errcodes-appendix.html)
+        if (err.code === '23505' && err.constraint === 'unique_group_name') {
+          throw new GroupOperationError('Group name is already taken.');
+        }
+        // Any other error is unexpected and should be handled by the main processes
+        throw err;
+      }
+      for (const uid of uids) {
+        await addUserToGroup({
+          assessment_id,
+          group_id,
+          uid,
+          authn_user_id,
+          enforceGroupSize: false,
+        });
+      }
     });
   } catch (err) {
-    flash(
-      'error',
-      `Failed to create the group ${groupName}. It is already taken. Please try another one.`,
-    );
+    if (err instanceof GroupOperationError) {
+      throw new GroupOperationError(`Failed to create the group ${group_name}. ${err.message}`);
+    }
+    throw err;
   }
+}
+
+export async function createOrAddToGroup(
+  group_name: string,
+  assessment_id: string,
+  uids: string[],
+  authn_user_id: string,
+): Promise<void> {
+  await sqldb.runInTransactionAsync(async () => {
+    const group = await sqldb.queryOptionalRow(
+      sql.select_and_lock_group_by_name,
+      { group_name, assessment_id },
+      GroupSchema,
+    );
+    if (group == null) {
+      await createGroup(group_name, assessment_id, uids, authn_user_id);
+    } else {
+      for (const uid of uids) {
+        await addUserToGroup({
+          assessment_id,
+          group_id: group.id,
+          uid,
+          authn_user_id,
+          enforceGroupSize: false,
+        });
+      }
+    }
+  });
 }
 
 /**
@@ -366,11 +488,11 @@ export async function leaveGroup(
       const currentSize = groupInfo.groupMembers.length;
       if (currentSize > 1) {
         const groupRoleAssignmentUpdates = getGroupRoleReassignmentsAfterLeave(groupInfo, userId);
-
-        await sqldb.queryAsync(sql.reassign_group_roles_after_leave, {
-          assessment_id: assessmentId,
+        await sqldb.queryAsync(sql.update_group_roles, {
           role_assignments: JSON.stringify(groupRoleAssignmentUpdates),
           group_id: groupId,
+          user_id: userId,
+          authn_user_id: authnUserId,
         });
 
         // Groups with low enough size should only use required roles
@@ -395,49 +517,54 @@ export async function leaveGroup(
   });
 }
 
-export async function getAssessmentPermissions(
-  assessmentId: string,
-  userId: string,
-): Promise<AssessmentLevelPermissions> {
-  return await sqldb.queryRow(
-    sql.get_assessment_level_permissions,
-    { assessment_id: assessmentId, user_id: userId },
-    AssessmentLevelPermissionsSchema,
-  );
+export function canUserAssignGroupRoles(groupInfo: GroupInfo, user_id: string): boolean {
+  const assignerRoles =
+    groupInfo.rolesInfo?.groupRoles
+      .filter((role) => role.can_assign_roles)
+      .map((role) => role.id) ?? [];
+  const assignerUsers = Object.values(groupInfo.rolesInfo?.roleAssignments ?? {})
+    .flat()
+    .filter((assignment) => assignerRoles.some((id) => idsEqual(id, assignment.group_role_id)))
+    .map((assignment) => assignment.user_id);
+  if (assignerUsers.length === 0) {
+    // If none of the current users in the group has an assigner role, allow any
+    // user to assign roles by default
+    return true;
+  } else {
+    // Otherwise, check if current user is in the list of assigner users
+    return assignerUsers.some((id) => idsEqual(id, user_id));
+  }
 }
 
 /**
- * Updates the role assignments of users in a group, given the output from groupRoleSelectTable.ejs.
+ * Updates the role assignments of users in a group, given the output from groupRoleTable.ejs.
  */
 export async function updateGroupRoles(
   requestBody: Record<string, any>,
   assessmentId: string,
+  groupId: string,
   userId: string,
+  hasStaffPermission: boolean,
   authnUserId: string,
 ) {
   await sqldb.runInTransactionAsync(async () => {
-    const groupId = await getGroupId(assessmentId, userId);
-    if (groupId === null) {
-      throw new Error(
-        "Couldn't access the user's group ID with the provided assessment and user IDs",
-      );
-    }
-
-    const permissions = await getAssessmentPermissions(assessmentId, userId);
-    if (!permissions.can_assign_roles_at_start) {
-      throw error.make(
-        403,
-        'User does not have permission to assign roles at the start of this assessment',
-      );
-    }
-
     const groupConfig = await getGroupConfig(assessmentId);
     const groupInfo = await getGroupInfo(groupId, groupConfig);
+
+    if (!hasStaffPermission && !canUserAssignGroupRoles(groupInfo, userId)) {
+      throw error.make(403, 'User does not have permission to assign roles');
+    }
 
     // Convert form data to valid input format for a SQL function
     const roleKeys = Object.keys(requestBody).filter((key) => key.startsWith('user_role_'));
     const roleAssignments = roleKeys.map((roleKey) => {
       const [roleId, userId] = roleKey.replace('user_role_', '').split('-');
+      if (!groupInfo.groupMembers.some((member) => idsEqual(member.user_id, userId))) {
+        throw error.make(403, `User ${userId} is not a member of this group`);
+      }
+      if (!groupInfo.rolesInfo?.groupRoles.some((role) => idsEqual(role.id, roleId))) {
+        throw error.make(403, `Role ${roleId} does not exist for this assessment`);
+      }
       return {
         group_id: groupId,
         user_id: userId,
@@ -448,12 +575,16 @@ export async function updateGroupRoles(
     // If no one is being given a role with assigner permissions, give that role to the current user
     const assignerRoleIds =
       groupInfo.rolesInfo?.groupRoles
-        .filter((role) => role.can_assign_roles_at_start)
+        .filter((role) => role.can_assign_roles)
         .map((role) => role.id) ?? [];
     const assignerRoleFound = roleAssignments.some((roleAssignment) =>
       assignerRoleIds.includes(roleAssignment.group_role_id),
     );
     if (!assignerRoleFound) {
+      if (!groupInfo.groupMembers?.some((member) => idsEqual(member.user_id, userId))) {
+        // If the current user is not in the group, this usually means they are a staff member, so give the assigner role to the first user
+        userId = groupInfo.groupMembers[0].user_id;
+      }
       roleAssignments.push({
         group_id: groupId,
         user_id: userId,
@@ -464,7 +595,6 @@ export async function updateGroupRoles(
     await sqldb.queryAsync(sql.update_group_roles, {
       group_id: groupId,
       role_assignments: JSON.stringify(roleAssignments),
-      user_id: userId,
       authn_user_id: authnUserId,
     });
   });