+++ apps/prairielearn/src/pages/courseSyncs/courseSyncs.js
@@ -137,10 +137,7 @@ router.post(
       const jobSequenceId = await syncHelpers.ecrUpdate(images, res.locals);
       res.redirect(res.locals.urlPrefix + '/jobSequence/' + jobSequenceId);
     } else {
-      throw error.make(400, 'unknown __action', {
-        locals: res.locals,
-        body: req.body,
-      });
+      throw error.make(400, `unknown __action: ${req.body.__action}`);
     }
   }),
 );