+++ apps/prairielearn/src/pages/instructorIssues/instructorIssues.js
@@ -1,19 +1,20 @@
 // @ts-check
 const asyncHandler = require('express-async-handler');
 const _ = require('lodash');
-const { parseISO, formatDistance } = require('date-fns');
-const express = require('express');
-const router = express.Router();
+import { parseISO, formatDistance } from 'date-fns';
+import * as express from 'express';
 const SearchString = require('search-string');
 const { z } = require('zod');
 
-const error = require('@prairielearn/error');
-const paginate = require('../../lib/paginate');
-const sqldb = require('@prairielearn/postgres');
-const { flash } = require('@prairielearn/flash');
-const { idsEqual } = require('../../lib/id');
-const { selectCourseInstancesWithStaffAccess } = require('../../models/course-instances');
+import * as error from '@prairielearn/error';
+import * as paginate from '../../lib/paginate';
+import * as sqldb from '@prairielearn/postgres';
+import { flash } from '@prairielearn/flash';
+import { idsEqual } from '../../lib/id';
+import { selectCourseInstancesWithStaffAccess } from '../../models/course-instances';
+import { closeAllIssuesForCourse } from '../../lib/issues';
 
+const router = express.Router();
 const sql = sqldb.loadSqlEquiv(__filename);
 
 const PAGE_SIZE = 100;
@@ -37,6 +38,9 @@ function formatForLikeClause(str) {
 
 function parseRawQuery(str) {
   const parsedQuery = SearchString.parse(str);
+  /**
+   * @type {{filter_is_open: boolean | null, filter_is_closed: boolean | null, filter_manually_reported: boolean | null, filter_automatically_reported: boolean | null, filter_qids: string[] | null, filter_not_qids: string[] | null, filter_query_text: string | null, filter_users: string[] | null, filter_not_users: string[] | null}}
+   */
   const filters = {
     /** @type {boolean | null} */
     filter_is_open: null,
@@ -226,7 +230,7 @@ router.get(
     res.locals.rows = issues.rows;
 
     res.locals.filterQuery = req.query.q;
-    res.locals.encodedFilterQuery = encodeURIComponent(/** @type {string } */ (req.query.q));
+    res.locals.encodedFilterQuery = encodeURIComponent((req.query.q ?? '').toString());
     res.locals.filters = filters;
     res.locals.anyFilteredIssuesOpen = issues.rows.some((row) => row.open);
 
@@ -274,14 +278,13 @@ router.post(
         z.number(),
       );
       flash('success', `Closed ${closedCount} ${closedCount === 1 ? 'issue' : 'issues'}.`);
+    } else if (req.body.__action === 'close_all') {
+      await closeAllIssuesForCourse(res.locals.course.id, res.locals.authn_user.user_id);
       res.redirect(req.originalUrl);
     } else {
-      throw error.make(400, 'unknown __action', {
-        locals: res.locals,
-        body: req.body,
-      });
+      throw error.make(400, `unknown __action: ${req.body.__action}`);
     }
   }),
 );
 
-module.exports = router;
+export default router;