+++ packages/aws/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/aws
 
+## 1.0.7
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
+## 1.0.6
+
+### Patch Changes
+
+- e53a188: Upgrade all JavaScript dependencies
+
+## 1.0.5
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+
 ## 1.0.4
 
 ### Patch Changes