+++ packages/config/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/config
 
+## 2.0.14
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+- Updated dependencies [dce0fa3]
+  - @prairielearn/aws-imds@1.0.13
+
+## 2.0.13
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+- Updated dependencies [abfd5cc]
+  - @prairielearn/aws-imds@1.0.12
+
 ## 2.0.12
 
 ### Patch Changes