+++ packages/flash/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/flash
 
+## 1.1.6
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+- Updated dependencies [dce0fa3]
+  - @prairielearn/html@3.1.4
+
+## 1.1.5
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+- Updated dependencies [abfd5cc]
+  - @prairielearn/html@3.1.3
+
 ## 1.1.4
 
 ### Patch Changes