+++ packages/html-ejs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @prairielearn/html-ejs
 
+## 1.1.16
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+- Updated dependencies [dce0fa3]
+  - @prairielearn/html@3.1.4
+
+## 1.1.15
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+- Updated dependencies [abfd5cc]
+  - @prairielearn/html@3.1.3
+
 ## 1.1.14
 
 ### Patch Changes