+++ packages/html/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/html
 
+## 3.1.4
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
+## 3.1.3
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+
 ## 3.1.2
 
 ### Patch Changes