+++ packages/postgres-tools/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @prairielearn/postgres-tools
 
+## 1.1.14
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+- Updated dependencies [dce0fa3]
+  - @prairielearn/postgres@1.9.1
+
+## 1.1.13
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+- Updated dependencies [abfd5cc]
+- Updated dependencies [3249e13]
+  - @prairielearn/postgres@1.9.0
+
 ## 1.1.12
 
 ### Patch Changes