+++ packages/sentry/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @prairielearn/sentry
 
+## 1.2.1
+
+### Patch Changes
+
+- dce0fa3: Upgrade all JavaScript dependencies
+
+## 1.2.0
+
+### Minor Changes
+
+- eebc7f9: Export `getCurrentScope` function
+
+## 1.1.5
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+
 ## 1.1.4
 
 ### Patch Changes